Refactoring: CLI and 4way-IF go through same code

+++ include/onewire.h
@@ -3,6 +3,6 @@
 extern void onewire_init();
 extern void onewire_putc(int c);
 extern void onewire_break();
-extern uint onewire_xfer(const void * tx_buf, uint tx_size, void * rx_buf, uint rx_size);
+extern size_t onewire_xfer(const void * tx_buf, size_t tx_size, void * rx_buf, size_t rx_size);
 extern void onewire_exit();
 #endif
